Fix warnings reported by clang -Weverything
- Remove sign warnings
- Add -Weverything to travis.yml
- Import portable uint64_t definition

+++ src/chars.h
@@ -2,13 +2,13 @@
 #define CHARS_H
 
 #include "inlines.h"
+#include "uu-stdint.h"
 
-struct SPDR_Chars
-{
-        int error;
+struct SPDR_Chars {
+        uint32_t error;
         char *chars;
-        int len;
-        int capacity;
+        size_t len;
+        size_t capacity;
 };
 
 #define NULL_CHARS                                                             \
